requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
chardet==4.0.0
#discord-components==2.1.2
git+https://github.com/kiki7000/discord.py-components.git@master
discord.py==1.7.3
idna==3.3
multidict==6.0.2
py-cord==1.7.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
toml==0.10.2
typing-extensions==4.3.0
yarl==1.7.2
